添加删除tag功能

+++ CocoaPodsTag/View/TagListView.swift
@@ -8,14 +8,22 @@
 import SwiftUI
 
 struct TagListView: View {
+    enum AlertType {
+        case normal, confirm
+    }
+    
+    @EnvironmentObject var store: Store
     @Environment(\.dismiss) var dismiss
     
     @State private var tagList: [String] = []
-    @State private var showAlert = false
-    @State private var alertMsg = ""
     @State private var selectedTag: String?
     
+    @State private var alertType: AlertType = .normal
+    @State private var showAlert = false
+    @State private var alertMsg = ""
+    
     @State private var showLoading = false
+    @State private var showSelectRemoteView = false
     
     var body: some View {
         List {
@@ -23,11 +31,14 @@ struct TagListView: View {
                 HStack {
                     Text(tag)
                     Spacer()
-//                    Button("删除") {
-//                        selectedTag = tag
-//                        showAlert = true
-//                        alertMsg = "删除tag:`\(tag)`"
-//                    }
+                    if !store.remotes.isEmpty {
+                        Button("删除") {
+                            selectedTag = tag
+                            showAlert = true
+                            alertType = .confirm
+                            alertMsg = "删除tag:`\(tag)`"
+                        }
+                    }
                 }
             }
         }
@@ -40,51 +51,74 @@ struct TagListView: View {
             }
         }
         .alert(isPresented: $showAlert) {
-            Alert(title: Text(alertMsg), message: nil, primaryButton: .destructive(Text("确认"), action: {
-                deleteTag()
-            }), secondaryButton: .default(Text("取消"), action: {
-                showAlert = false
-            }))
+            alertView
         }
         .sheet(isPresented: $showLoading, onDismiss: {
             showLoading = false
         }, content: {
             LoadingView()
         })
+        .sheet(isPresented: $showSelectRemoteView, onDismiss: {
+            showSelectRemoteView = false
+        }, content: {
+            SelectRemoteView(confirmAction: {
+                deleteTag()
+            })
+        })
         .onAppear {
             loadTagList()
         }
     }
     
+    private var alertView: Alert {
+        if alertType == .normal {
+            return Alert(title: Text(alertMsg), message: nil, dismissButton: .default(Text("OK")))
+        } else {
+            return Alert(title: Text(alertMsg), message: nil, primaryButton: .destructive(Text("确认"), action: {
+                deleteTag()
+            }), secondaryButton: .default(Text("取消"), action: {
+                showAlert = false
+            }))
+        }
+    }
+    
     private func deleteTag() {
         guard let tag = selectedTag else { return }
+        guard let remote = store.remote else {
+            showSelectRemoteView = true
+            return
+        }
         showLoading = true
         DispatchQueue.global().async {
-            let error = Command.sharedInstance.deleteLocalTag(tag)
+            let error = Command.sharedInstance.deleteRemoteTag(tag, remote: remote)
             DispatchQueue.main.async {
                 if error == nil {
-                    deleteRemoteTag()
+                    debugPrint("删除远端tag:`\(tag)`成功")
+                    deleteLocalTag(tag)
                 } else {
+                    showLoading = false
                     showAlert = true
-                    alertMsg = "删除本地tag失败\n\(error!)"
+                    alertType = .normal
+                    alertMsg = "删除远端tag失败\n\(error!)"
                 }
             }
         }
     }
     
-    private func deleteRemoteTag() {
-        guard let tag = selectedTag else { return }
+    private func deleteLocalTag(_ tag: String) {
         showLoading = true
         DispatchQueue.global().async {
-            let error = Command.sharedInstance.deleteRemoteTag(tag, remote: "origin")
+            let error = Command.sharedInstance.deleteLocalTag(tag)
             DispatchQueue.main.async {
                 selectedTag = nil
                 showLoading = false
                 if error == nil {
+                    debugPrint("删除本地tag:`\(tag)`成功")
                     loadTagList()
                 } else {
                     showAlert = true
-                    alertMsg = "删除远程tag失败\n\(error!)"
+                    alertType = .normal
+                    alertMsg = "删除本地tag失败\n\(error!)"
                 }
             }
         }
